pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mf"
version = "0.1.0"
description = "Metafunctor site management tools - papers, projects, and publications"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Alex Towell", email = "lex@metafunctor.com" }
]
keywords = ["hugo", "static-site", "papers", "github", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
pdf = [
    "pypdf>=3.0",
    "pdf2image>=1.16",
    "Pillow>=9.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
all = [
    "mf[pdf,dev]",
]

[project.scripts]
mf = "mf.cli:main"

[project.urls]
Homepage = "https://metafunctor.com"
Repository = "https://github.com/queelius/metafunctor"

[tool.hatch.build.targets.wheel]
packages = ["src/mf"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

# Include markdown files in claude/data/
[tool.hatch.build]
include = [
    "src/mf/**/*.py",
    "src/mf/claude/data/*.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/mf"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
